requirements.txt
# pip install -r requirements.txt

# Base ----------------------------------------
gitpython
ipython>=8.6.0
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
psutil  # system resources
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
torch>=1.7.0  # see https://pytorch.org/get-started/locally (recommended)
torchvision>=0.8.1
tqdm>=4.64.0
# protobuf<=3.20.1  # https://github.com/ultralytics/yolov5/issues/8012

# Logging ---------------------------------------------------------------------
tensorboard>=2.4.1
# clearml>=1.2.0
# comet

# Visualizing --------------------------------------------------------------------
pandas>=1.1.4
seaborn>=0.11.0
plotly>=5.11.0
ffmpegio>=0.7.0

# StrongSORT ------------------------------------------------------------------
easydict>=1.10

# torchreid -------------------------------------------------------------------
gdown>=4.5.3

# ByteTrack -------------------------------------------------------------------
lap>=0.4.0
git+https://github.com/samson-wang/cython_bbox.git@9badb346a9222c98f828ba45c63fe3b7f2790ea2

# OCSORT ----------------------------------------------------------------------
filterpy>=1.4.5

# Inference ----------------------------------------------------------------------
# onnx>=1.9.0  # ONNX
# onnx-simplifier>=0.4.1  # ONNX
# onnxruntime>=1.13.1 # ONNX
# nvidia-pyindex  # TensorRT
# nvidia-tensorrt  # TensorRT
#  pycuda>=2022.2